+++ auto_route/lib/src/router/controller/routing_controller.dart
@@ -11,6 +11,7 @@ import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
+import 'package:path/path.dart' as p;
 
 part '../../route/route_data.dart';
 part 'auto_route_guard.dart';