+++ auto_route_generator/lib/export_lib/exporte_1.dart
@@ -1,6 +1,0 @@
-export '../builder.dart';
-export '../src/models/resolved_type.dart';
-
-export 'package:code_builder/code_builder.dart';
-
-class Widget  {}