+++ auto_route_generator/lib/export_lib/feature.dart
@@ -1,5 +1,0 @@
-import 'exporte_1.dart';
-// import 'package:flutter/material.dart' show Widget;
-
-// late final Block block;
-